edited count

+++ wda-persister/src/main/java/ch/hslu/informatik/swde/wda/persister/DAO/WeatherDAO.java
@@ -21,7 +21,7 @@ public interface WeatherDAO extends GenericDAO<Weather> {
      *
      * @return Anzahl der Wetterdaten pro Stadt
      */
-    int getNumberOfWeatherByCity(String cityName);
+    long getNumberOfWeatherByCity(String cityName);
 
     /**
      * Holt das neueste Weather für eine bestimmte Ortschaft basierend auf der Ortschafts-ID.
